refactor(app): split view rendering helpers
